remove global concept of MessageTypes. All tests are adjusted accordingly.
