+++ cardinal/gamestate/tick_test.go
@@ -2,11 +2,11 @@ package gamestate_test
 
 import (
 	"context"
-	"pkg.world.dev/world-engine/cardinal/message"
+	"testing"
+
 	"pkg.world.dev/world-engine/cardinal/testutils"
 	"pkg.world.dev/world-engine/cardinal/types"
 	"pkg.world.dev/world-engine/cardinal/types/txpool"
-	"testing"
 
 	"pkg.world.dev/world-engine/assert"
 )
@@ -19,8 +19,8 @@ func TestCanSaveAndRecoverTransactions(t *testing.T) {
 		Value int
 	}
 
-	msgAlpha := message.NewMessageType[MsgIn, MsgOut]("alpha")
-	msgBeta := message.NewMessageType[MsgIn, MsgOut]("beta")
+	msgAlpha := testutils.NewMessageType[MsgIn, MsgOut]("alpha")
+	msgBeta := testutils.NewMessageType[MsgIn, MsgOut]("beta")
 	assert.NilError(t, msgAlpha.SetID(16))
 	assert.NilError(t, msgBeta.SetID(32))
 	msgs := []types.Message{msgAlpha, msgBeta}
