+++ cardinal/message/message.go
@@ -40,10 +40,10 @@ func isStruct[T any]() bool {
 		inKind == reflect.Struct
 }
 
-// NewMessageType creates a new message type. It accepts two generic type parameters: the first for the message input,
+// newMessageType creates a new message type. It accepts two generic type parameters: the first for the message input,
 // which defines the data needed to make a state transition, and the second for the message output, commonly used
 // for the results of a state transition.
-func NewMessageType[In, Out any](
+func newMessageType[In, Out any](
 	name string,
 	opts ...MessageOption[In, Out],
 ) *MessageType[In, Out] {
@@ -61,6 +61,14 @@ func NewMessageType[In, Out any](
 		opt(msg)
 	}
 	return msg
+}
+
+func (t *MessageType[In, Out]) SetName(name string) {
+	t.name = name
+}
+
+func (t *MessageType[In, Out]) SetGroup(group string) {
+	t.group = group
 }
 
 func (t *MessageType[In, Out]) Name() string {
@@ -179,7 +187,7 @@ func (t *MessageType[In, Out]) ABIEncode(v any) ([]byte, error) {
 
 	var args ethereumAbi.Arguments
 	var input any
-	//nolint:gocritic // its fine.
+	//nolint:gocritic // it's fine.
 	switch in := v.(type) {
 	case Out:
 		input = in
