+++ cardinal/plugin_persona.go
@@ -14,6 +14,7 @@ import (
 	"pkg.world.dev/world-engine/cardinal/persona/query"
 	"pkg.world.dev/world-engine/cardinal/types"
 	"pkg.world.dev/world-engine/cardinal/types/engine"
+	"pkg.world.dev/world-engine/cardinal/worldstage"
 )
 
 type personaPlugin struct {
@@ -72,27 +73,30 @@ func (p *personaPlugin) RegisterComponents(world *World) error {
 }
 
 func (p *personaPlugin) RegisterMessages(world *World) error {
-	err := RegisterMessages(world, CreatePersonaMsg, AuthorizePersonaAddressMsg)
-	if err != nil {
-		return err
+	if world.worldStage.Current() != worldstage.Init {
+		return eris.Errorf(
+			"engine state is %s, expected %s to register messages",
+			world.worldStage.Current(),
+			worldstage.Init,
+		)
 	}
-	return nil
+	return errors.Join(
+		RegisterMessage[msg.CreatePersona, msg.CreatePersonaResult](
+			world,
+			"create-persona",
+			message.WithCustomMessageGroup[msg.CreatePersona, msg.CreatePersonaResult]("persona"),
+			message.WithMsgEVMSupport[msg.CreatePersona, msg.CreatePersonaResult]()),
+		RegisterMessage[msg.AuthorizePersonaAddress, msg.AuthorizePersonaAddressResult](
+			world,
+			"authorize-persona-address",
+		))
+	//return registerMessagesByName(world, CreatePersonaMsg, AuthorizePersonaAddressMsg)
+
 }
 
 // -----------------------------------------------------------------------------
 // Persona Messages
 // -----------------------------------------------------------------------------
-
-var AuthorizePersonaAddressMsg = message.NewMessageType[msg.AuthorizePersonaAddress, msg.AuthorizePersonaAddressResult](
-	"authorize-persona-address",
-)
-
-// CreatePersonaMsg is a message that facilitates the creation of a persona tag.
-var CreatePersonaMsg = message.NewMessageType[msg.CreatePersona, msg.CreatePersonaResult](
-	"create-persona",
-	message.WithCustomMessageGroup[msg.CreatePersona, msg.CreatePersonaResult]("persona"),
-	message.WithMsgEVMSupport[msg.CreatePersona, msg.CreatePersonaResult](),
-)
 
 // AuthorizePersonaAddressSystem enables users to authorize an address to a persona tag. This is mostly used so that
 // users who want to interact with the game via smart contract can link their EVM address to their persona tag, enabling
@@ -102,8 +106,11 @@ func AuthorizePersonaAddressSystem(wCtx engine.Context) error {
 	if err != nil {
 		return err
 	}
-
-	AuthorizePersonaAddressMsg.Each(
+	authorizePersonaAddressMsg, err := GetMessage[msg.AuthorizePersonaAddress, msg.AuthorizePersonaAddressResult](wCtx)
+	if err != nil {
+		return err
+	}
+	authorizePersonaAddressMsg.Each(
 		wCtx,
 		func(txData message.TxData[msg.AuthorizePersonaAddress]) (
 			result msg.AuthorizePersonaAddressResult, err error,
@@ -158,8 +165,11 @@ func RegisterPersonaSystem(wCtx engine.Context) error {
 	if err != nil {
 		return err
 	}
-
-	CreatePersonaMsg.Each(
+	createPersonaMsg, err := GetMessage[msg.CreatePersona, msg.CreatePersonaResult](wCtx)
+	if err != nil {
+		return err
+	}
+	createPersonaMsg.Each(
 		wCtx,
 		func(txData message.TxData[msg.CreatePersona]) (result msg.CreatePersonaResult, err error) {
 			txMsg := txData.Msg
