+++ cardinal/router/iterator/iterator_test.go
@@ -4,15 +4,17 @@ import (
 	"context"
 	"encoding/binary"
 	"errors"
+	"testing"
+
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/proto"
+
 	"pkg.world.dev/world-engine/assert"
-	"pkg.world.dev/world-engine/cardinal/message"
 	"pkg.world.dev/world-engine/cardinal/router/iterator"
+	"pkg.world.dev/world-engine/cardinal/testutils"
 	"pkg.world.dev/world-engine/cardinal/types"
 	shardtypes "pkg.world.dev/world-engine/evm/x/shard/types"
 	shard "pkg.world.dev/world-engine/rift/shard/v2"
-	"testing"
 )
 
 var _ shardtypes.QueryClient = &mockQuerier{}
@@ -75,7 +77,7 @@ func TestIteratorReturnsErrorIfQueryFails(t *testing.T) {
 type fooIn struct{ X int }
 type fooOut struct{}
 
-var fooMsg = message.NewMessageType[fooIn, fooOut]("foo")
+var fooMsg = testutils.NewMessageType[fooIn, fooOut]("foo")
 
 func TestIteratorHappyPath(t *testing.T) {
 	err := fooMsg.SetID(10)
