+++ cardinal/server/server_test.go
@@ -30,6 +30,24 @@ import (
 	"pkg.world.dev/world-engine/sign"
 )
 
+// Used for Registering message
+type MoveMsgInput struct {
+	Direction string
+}
+
+// Used for Registering message
+type MoveMessageOutput struct {
+	Location LocationComponent
+}
+
+type QueryLocationRequest struct {
+	Persona string
+}
+
+type QueryLocationResponse struct {
+	LocationComponent
+}
+
 type ServerTestSuite struct {
 	suite.Suite
 
@@ -63,10 +81,12 @@ func (s *ServerTestSuite) TestCanClaimPersonaSendGameTxAndQueryGame() {
 	s.setupWorld()
 	s.fixture.DoTick()
 	personaTag := s.CreateRandomPersona()
-	s.runTx(personaTag, MoveMessage, MoveMsgInput{Direction: "up"})
+	moveMessage, err := cardinal.GetMessageFromWorld[MoveMsgInput, MoveMessageOutput](s.world)
+	s.Require().NoError(err)
+	s.runTx(personaTag, moveMessage, MoveMsgInput{Direction: "up"})
 	res := s.fixture.Post("query/game/location", QueryLocationRequest{Persona: personaTag})
 	var loc LocationComponent
-	err := json.Unmarshal([]byte(s.readBody(res.Body)), &loc)
+	err = json.Unmarshal([]byte(s.readBody(res.Body)), &loc)
 	s.Require().NoError(err)
 	s.Require().Equal(loc, LocationComponent{0, 1})
 }
@@ -145,7 +165,9 @@ func (s *ServerTestSuite) TestCanSendTxWithoutSigVerification() {
 		PersonaTag: persona,
 		Body:       msgBz,
 	}
-	url := "/tx/game/" + MoveMessage.Name()
+	moveMessage, err := cardinal.GetMessageFromWorld[MoveMsgInput, MoveMessageOutput](s.world)
+	s.Require().NoError(err)
+	url := "/tx/game/" + moveMessage.Name()
 	res := s.fixture.Post(url, tx)
 	s.Require().Equal(fiber.StatusOK, res.StatusCode, s.readBody(res.Body))
 	err = s.world.Tick(context.Background(), uint64(time.Now().Unix()))
@@ -215,11 +237,15 @@ func (s *ServerTestSuite) setupWorld(opts ...cardinal.WorldOption) {
 	s.world = s.fixture.World
 	err := cardinal.RegisterComponent[LocationComponent](s.world)
 	s.Require().NoError(err)
-	err = cardinal.RegisterMessages(s.world, MoveMessage)
+	err = cardinal.RegisterMessage[MoveMsgInput, MoveMessageOutput](s.world, "move")
 	s.Require().NoError(err)
 	personaToPosition := make(map[string]types.EntityID)
 	err = cardinal.RegisterSystems(s.world, func(context engine.Context) error {
-		MoveMessage.Each(context, func(tx message.TxData[MoveMsgInput]) (MoveMessageOutput, error) {
+		moveMessage, err := cardinal.GetMessage[MoveMsgInput, MoveMessageOutput](context)
+		if err != nil {
+			return err
+		}
+		moveMessage.Each(context, func(tx message.TxData[MoveMsgInput]) (MoveMessageOutput, error) {
 			posID, exists := personaToPosition[tx.Tx.PersonaTag]
 			if !exists {
 				id, err := cardinal.Create(context, LocationComponent{})
@@ -293,22 +319,4 @@ type LocationComponent struct {
 
 func (LocationComponent) Name() string {
 	return "location"
-}
-
-type MoveMsgInput struct {
-	Direction string
-}
-
-type MoveMessageOutput struct {
-	Location LocationComponent
-}
-
-var MoveMessage = message.NewMessageType[MoveMsgInput, MoveMessageOutput]("move")
-
-type QueryLocationRequest struct {
-	Persona string
-}
-
-type QueryLocationResponse struct {
-	LocationComponent
 }
