+++ cardinal/state_test.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"testing"
 
+	"time"
+
 	"pkg.world.dev/world-engine/cardinal"
 	"pkg.world.dev/world-engine/cardinal/filter"
 	"pkg.world.dev/world-engine/cardinal/iterators"
-	"pkg.world.dev/world-engine/cardinal/message"
 	"pkg.world.dev/world-engine/cardinal/types"
 	"pkg.world.dev/world-engine/cardinal/types/engine"
-	"time"
 
 	"pkg.world.dev/world-engine/assert"
 
@@ -302,7 +302,7 @@ func TestEngineTickAndHistoryTickMatch(t *testing.T) {
 func TestCanFindTransactionsAfterReloadingEngine(t *testing.T) {
 	type Msg struct{}
 	type Result struct{}
-	someTx := message.NewMessageType[Msg, Result]("some-msg")
+	//someTx := message.NewMessageType[Msg, Result]("some-msg")
 	redisStore := miniredis.RunT(t)
 	ctx := context.Background()
 
@@ -311,10 +311,14 @@ func TestCanFindTransactionsAfterReloadingEngine(t *testing.T) {
 	for reload := 0; reload < 5; reload++ {
 		tf := testutils.NewTestFixture(t, redisStore)
 		world := tf.World
-		assert.NilError(t, cardinal.RegisterMessages(world, someTx))
+		assert.NilError(t, cardinal.RegisterMessage[Msg, Result](world, "some-msg"))
 		err := cardinal.RegisterSystems(
 			world,
 			func(wCtx engine.Context) error {
+				someTx, err := cardinal.GetMessage[Msg, Result](wCtx)
+				if err != nil {
+					return err
+				}
 				for _, tx := range someTx.In(wCtx) {
 					someTx.SetResult(wCtx, tx.Hash, Result{})
 				}
@@ -325,6 +329,7 @@ func TestCanFindTransactionsAfterReloadingEngine(t *testing.T) {
 		tf.StartWorld()
 
 		relevantTick := world.CurrentTick()
+		someTx, err := cardinal.GetMessageFromWorld[Msg, Result](world)
 		for i := 0; i < 3; i++ {
 			_ = tf.AddTransaction(someTx.ID(), Msg{}, testutils.UniqueSignature())
 		}
