+++ cardinal/world.go
@@ -201,10 +201,6 @@ func NewMockWorld(opts ...WorldOption) (*World, error) {
 	return world, nil
 }
 
-func (w *World) registerMessagesByName(msgs ...types.Message) error {
-	return w.msgManager.RegisterMessages(msgs...)
-}
-
 func GetMessageFromWorld[In any, Out any](world *World) (*message.MessageType[In, Out], error) {
 	var msg message.MessageType[In, Out]
 	msgType := reflect.TypeOf(msg)
@@ -320,7 +316,7 @@ func (w *World) GetMessageManager() *message.Manager {
 
 // StartGame starts running the world game loop. Each time a message arrives on the tickChannel, a world tick is
 // attempted. In addition, an HTTP server (listening on the given port) is created so that game messages can be sent
-// to this world. After StartGame is called, RegisterComponent, RegisterMessages,
+// to this world. After StartGame is called, RegisterComponent, registerMessagesByName,
 // RegisterQueries, and RegisterSystems may not be called. If StartGame doesn't encounter any errors, it will
 // block forever, running the server and ticking the game in the background.
 func (w *World) StartGame() error {
